+++ src/modules/sound/lullaby/FLACDecoder.h
@@ -46,7 +46,7 @@ public:
 	static bool accepts(const std::string &ext);
 	love::sound::Decoder *clone();
 	int decode();
-	bool seek(float s);
+	bool seek(double s);
 	bool rewind();
 	bool isSeekable();
 	int getChannelCount() const;
